renamed boardGetters
