+++ core/src/game/state.ts
@@ -1,4 +1,4 @@
-import { getLastPlayedTile } from "../board/boardGetters";
+import { getLastPlayedTile } from "../board/boardService";
 import { Board } from "../board/boardType";
 import { getPlayableTilesForNextMove } from "../move/move";
 import { BLACK_PLAYER, Player } from "../player/player";
