+++ core/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./board/board";
-export * from "./board/boardGetters";
+export * from "./board/boardService";
 export * from "./board/boardType";
 export * from "./game/state";
 export * from "./graph/graph";
